move native key material
