+++ crates/iceberg/src/encryption/encrypted_io.rs
@@ -21,9 +21,9 @@ use std::sync::Arc;
 
 use bytes::Bytes;
 
+use super::crypto::SensitiveBytes;
 use super::file_decryptor::AesGcmFileDecryptor;
 use super::file_encryptor::AesGcmFileEncryptor;
-use super::key_metadata::NativeKeyMaterial;
 use crate::io::{FileMetadata, FileRead, FileWrite, InputFile, OutputFile};
 
 /// An AGS1 stream-encrypted input file wrapping a plain [`InputFile`].
@@ -246,5 +246,44 @@ impl std::fmt::Debug for NativeEncryptedOutputFile {
         f.debug_struct("NativeEncryptedOutputFile")
             .field("path", &self.inner.location())
             .finish_non_exhaustive()
+    }
+}
+
+/// Plaintext key material for Parquet Modular Encryption (PME).
+///
+/// Carries the DEK and AAD prefix needed by a Parquet reader/writer to
+/// configure `FileEncryptionProperties` / `FileDecryptionProperties`.
+///
+/// Rust equivalent of Java's `NativeEncryptionKeyMetadata` interface.
+pub struct NativeKeyMaterial {
+    dek: SensitiveBytes,
+    /// AAD prefix is not secret — it is stored in plaintext in file metadata
+    /// and used for integrity (authenticated data), not confidentiality.
+    aad_prefix: Box<[u8]>,
+}
+
+impl NativeKeyMaterial {
+    /// Creates a new `NativeKeyMaterial`.
+    pub fn new(dek: SensitiveBytes, aad_prefix: Box<[u8]>) -> Self {
+        Self { dek, aad_prefix }
+    }
+
+    /// Returns the plaintext DEK.
+    pub fn dek(&self) -> &SensitiveBytes {
+        &self.dek
+    }
+
+    /// Returns the AAD prefix.
+    pub fn aad_prefix(&self) -> &[u8] {
+        &self.aad_prefix
+    }
+}
+
+impl std::fmt::Debug for NativeKeyMaterial {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("NativeKeyMaterial")
+            .field("dek", &self.dek)
+            .field("aad_prefix_len", &self.aad_prefix.len())
+            .finish()
     }
 }
