+++ crates/iceberg/src/encryption/mod.rs
@@ -32,10 +32,11 @@ mod stream;
 pub use crypto::{AesGcmCipher, AesKeySize, SecureKey, SensitiveBytes};
 pub use encrypted_io::{
     EncryptedInputFile, EncryptedOutputFile, NativeEncryptedInputFile, NativeEncryptedOutputFile,
+    NativeKeyMaterial,
 };
 pub use encryption_manager::EncryptionManager;
 pub use file_decryptor::AesGcmFileDecryptor;
 pub use file_encryptor::AesGcmFileEncryptor;
-pub use key_metadata::{NativeKeyMaterial, StandardKeyMetadata};
+pub use key_metadata::StandardKeyMetadata;
 pub use kms::{GeneratedKey, KeyManagementClient};
 pub use stream::{AesGcmFileRead, AesGcmFileWrite};
